Separa componentes e telas
