+++ dojos/20170927-react-native/demo/screens/Details.js
@@ -1,0 +1,25 @@
+import React from 'react';
+import {
+  Button,
+  Image,
+  StyleSheet,
+  Text,
+  View,
+} from 'react-native';
+
+class Details extends React.Component {
+  static navigationOptions = {
+    title: 'Detalhes',
+  };
+
+  render() {
+    const { params } = this.props.navigation.state;
+    return (
+      <View>
+        <Text>Hospedagem id: {params.id}</Text>
+      </View>
+    );
+  }
+}
+
+export default Details;
