+++ dojos/20170927-react-native/demo/screens/Gallery.js
@@ -1,0 +1,45 @@
+import React from 'react';
+import {
+  Button,
+  Image,
+  StyleSheet,
+  Text,
+  View,
+} from 'react-native';
+
+import Place from '../components/Place';
+
+class Gallery extends React.Component {
+  static navigationOptions = {
+    title: 'Hospedagens',
+  };
+
+  state = {
+    places: [],
+  }
+
+  componentDidMount() {
+    fetch('http://localhost:7070/places')
+    .then((res) => res.json())
+    .then(this.updatePlacesList);
+  }
+
+  updatePlacesList = (places) => {
+    this.setState({
+      places
+    });
+  }
+
+  render() {
+    const { navigate } = this.props.navigation;
+    return (
+      <View>
+        {this.state.places.map((place) => (
+          <Place key={place.id} data={place} navigate={navigate}></Place>
+        ))}
+      </View>
+    );
+  }
+}
+
+export default Gallery;
